[CDEC-416] Add a new sinbin package
This is a temporary holding area for code that needs to be moved while
the package dependency graph is reorganised.

Specifically, code that is in `infra` that the blockchain packages
(`block`, `txp`, `delegation` etc) depend with be moved to this
package so that eventually `infra` can depend on the blockchain
packages.